Change map/array indexing syntax.
Rather than having specific tokenization for ":" and specific parsing
for indexing, this opts for using command expressions to index into
arrays and maps.

It also allows chained commands; when the applicant of the command is an
array or map, the next argument is used to index into the value, but
the remaining arguments are applied as if they were a command to the
resulting value after indexing. The only exception here is that string
environment lookup only occurs for the very first value; if an array or
map value evaluates to a string, it will not be looked up in the
environment and will be considered as a string.

In this commit, the application command-line behavior is also changed.
Previously, each additional argument would be looked up separately in a
map resulting from script execution. Now, the command line aligns
exactly with the script. If one types '(binary) a b c', it's as if one
executed a script with those lines (i.e., it uses the loading builtin to
load the file 'a' and pass the remaining arguments ('b' and 'c') to the
resulting value from executing the file.