Remove `do` in favor of `seq`.
`seq` runs arguments sequentially, returning a value with the last
value result.
Previously `do` suffered from work logging being incorrect, because it
would start logging when a Value was first run. Now, work logging is
done existentially, and work is tied to Value future lifetimes. This
required a few changes throughout the code, especially a bugfix in
`exec` to discard run_command when it wouldn't be needed.