Change parsing as a result of bind syntax RFC.
For functions, expressions, set expressions, etc. the syntax now does
not require parentheses for commands. If more than one argument is
supplied, a command is called. If only one argument is supplied and it
is a string, a literal string is returned. Parentheses or a dollar sign
can be used to force a command with a single string literal.