Many changes to simplify runtime.
These changes start to expose simple operators to scripts so they may
compose and adjust things as they see fit.

The `cache` command has been added, which includes a new Stored trait in
the grease runtime. This trait specifies how to store and load arbitrary
values in grease store items. String and PathBuf are currently
implemented (but many more can be added).

The `variable` experimental command has been added, which allows a user
to clear the dependencies of arbitrary values (thus making the
identifier consistent), optionally specifying a separate value to set as
a dependency (making the identifier consistent with a given value). This
can be used to, for instance, ignore changes of a value.

The `path` command has been changed to a map (module), with three
commands:
* `new` to create a random path,
* `join` to join path components into a path, and
* `split` to split a path into an array of components.

Since `new` creates a random path, its identity is constant (the random
path it creates is not part of the dependency chain; the _fact_ that it
creates a random path is all that is important).