Change pattern matching destructuring to produce values which rely on original values.
This change makes it easy to write correct (efficient) scripts which
have values that change id only as required. Previously destructuring
would use the original value ids of nested values, however it is more
correct to produce new ids which capture the fact that the value relies
on the result of destructuring.