Improvements to codebase, add Nested grease trait, support Stored grease trait for all script types.
The grease `Context<T>` nested type was renamed from `Context` to
`Runtime` to avoid silly conflicts and confusion.

The `Nested` grease trait is used to expose nested `Value`s within the
resulting data of a `Value`. It is now used to force values recursively
to evaluate, removing the old, non-type-variant version.

The `Nested` trait was necessary to generalize over arbitrary types and
be able to force values in an async context such that later in a
non-async context (such as serializing nested `Value`s in a `Store`
implementation) you can peek to get the values.

The `Stored` trait has now been implemented for all script types
(`ScriptUnit`, `ScriptString`, `ScriptArray`, `ScriptMap`) and
`PathBuf`.