Add support for prelude modules.
If a prelude file is available in the load path, it will automatically
be loaded and merged into the script environment. It must evaluate to a
script map.

This also fixes an issue where script environments would have calling
script environment values present. Now environments are cleared
appropriately when scripts are loaded.