Plain strings within blocks are now shorthand for binding.
For example, `{a,b}` is shorthand for `{a = $a,b = $b}`.
This change removes no features while making map creation more
convenient. It removes no features because a string as a line in a block
would otherwise be completely pointless: there are no side effects and
the value is not used.