Fill out the essential functions in the standard module.
Change the semantics of exec to not spawn in a new task.
Add a standard module `task` function to run things in a separate task.
Alter the output format for tasks so that they only display those that
are running, if any.

More logging needs to be added, as well as hooks for users to log and
choose units of work.