Change how workspaces are accessed.
The prelude is loaded by applying 'prelude' to the evaluated workspace
value of the exact located workspace (rather than piggybacking off of
fallback behavior).

The fallback behavior is changed to explicitly apply 'command' to the
evaluated workspace value of the located ancestor workspace, and _then_
applying any remaining arguments if they exist. If no arguments exist,
it returns the value as-is.

If a script is loaded directly (not from fallback behavior in a
workspace) and evaluates to a function, that function is always applied,
even if there are no additional arguments.

* Eager

To unify function and array/map application, an Eager type was added to
grease and threaded through grease values. The Eager type interoperates
with the Value's internal shared futures and allows value adaptation to
occur eagerly when possible. It also propagates eagerness through these
adaptations such that resulting shared futures within values have
knowledge of whether they have already been eagerly evaluated. This is
essential for function application to occur ASAP, but also has a nice
side effect of eagerly applying array/map indices to get
(possibly-typed) values immediately if the array/map is immediately
available.