Add a copy_interactive io function for quick flushing.
This is probably not necessary at all, but was the result of trying to
debug why ssh wouldn't work correctly. For the record, it is because ssh
reads/writes directly with `/dev/tty` when getting a password and had
nothing to do with the io code.

Update project workspace to deploy mac package.