Rework how tasks are tracked.
The semantics of tasks are now the following:

If a task is waiting on another task, it is considered inactive (and
task logging reflects this). It cedes its task-count for other tasks to
run (which also prevents any possibility of deadlock, since the
task-count is always at most the maximum tasks allowed).