Improve error tracking by using task-local variables.
Errors across task boundaries are fanned out to all values that await
the task value.

Errors in `error:catch` are ignored.

Errors are better deduplicated based on their source references.

To facilitate early shutdown (and ease the burden of cleanup), loaded
plugins are now leaked rather than cleaned up in the context. This could
be avoided by using weak references to the tokio runtime in contexts (so
that any values that use the grease runtime don't keep the tokio runtime
alive). Then the main thread would control the lifetime and shutdown of
the tokio runtime (which might be for the best).

Change `seq` to correctly obey the `-S` flag (i.e. aggregate errors or
return immediately).