Make `ergo` (load) more consistent.
Do not apply functions without arguments if loaded in scripts.

The program will force the final value, including applying functions if
the final value is a function (with no arguments).

This makes the behavior of `ergo` in scripts less confusing. It will
always return a value if called without additional arguments.