Improve Value::for_each. Fix task behavior.
Change how active task tracking works to make task persistence opt-in
(explicit) rather than implicit from handoff.

Unfortunately due to the way that shared futures are used in Values, it
is not possible to determine whether a future may effectively await a
task or not (when one is already running), so unless Shared futures are
changed to propagate this information, it is not possible to handoff
tasks automatically. Task errors are also subject to this issue, but
that is of much lesser consequence than the task running as improper
task management easily causes deadlocks.