Major rewrite of parsing and the runtime.
A lot of things were able to be removed and simplified greatly. See the
ChangeLog for a more detailed explanation of the syntax changes.

Bind replaces Call as the defacto trait for delayed computation. It is
generalized to take a single value, which may be a group of call
arguments, bind arguments, or anything else. In this way a single value
can behave differently depending on the value being bound and/or the
expression context.