Change parentheses to be purely grouping. Make indexing a distinct operation (as opposed to being syntax sugar).
Previously, parentheses implied commands. Now they are strictly for
grouping, e.g. `(a)` and `a` parse identically. Commands are inferred
based on the presence of arguments. To call a command with no arguments,
use the trailing colon, e.g. `a:`.

Indexing (infix `:`) is now a distinct operation, binding an `Index`
value. In scripts, you may match this with `index :v`. Maps and arrays
have been changed to use `Index` rather than `Args` to get values.
Syntactically this only has breaking implications if a script was
retrieving map/array values using a function call (`map :key`), or were
using `:` on things that were not maps or arrays (`function:arg`).