Refactor and improve the script runtime.
This is a major refactor and cleanup of ergo_runtime. The grease crate
is removed; parts of it are retained in abi_stable_ext, whereas other
parts were moved to ergo_runtime. ergo_runtime has been reorganized and
cleaned up as well. ergo_script has been updated to evaluate values
based on this new runtime, and all tests are passing.

The runtime has had a lot of cruft removed, and Values have been
simplified such that they are either dynamically-typed (and hold a
future to be evaluated) or are typed and their data is available
immediately. This allows errors to be lifted into Values, such that any
errors are simply a script `Error` type. Values are also given the
context at their (first) evaluation site.