Finish up most of the scripted side of ergo_std.
There were a number of bugs to fix and design decisions to make in the
runtime. With the new evaluation semantics, it becomes
easy/clean/natural to simply use the presence of any error as an
indication that a match call should try to proceed to the next case.
This simplification allowed us to remove the PatternError class entirely
(which was always dubious). Now things are much more predictable.

Fixed the docstring generation for `ergo_fn` et al macros; didn't
realize that there is a doc attribute for each line (and needed to do a
little cleaning of the strings).