Correct how dynamic binding access is detected.
The runtime Context can no longer be cloned, it can only be forked and
joined. This is essential as joining allows flags/information set in a
forked (similar to cloned) child to be accounted for and merged into the
parent context.

The use of Context has been pervasive enough that it may end up being
appropriate to make it a task-local variable.