Only force indexed expressions when the value being indexed is a capture.
This is generally the only case where you'd want this to be forced, and
makes things behave in a more intuitive manner.

We may want to extend somewhat related semantics to functions; e.g. if
any binding introduced by a function is used in a capture within the
function, any bind sites of the function will be implicitly forced.