Change capture evaluation in release builds.
In debug builds, captures _must_ exist in the capture scope and pattern
commands evaluate their arguments deeply.

As a small optimization, in release builds we assume that if a capture
is not in the capture scope it was never set (but could have been), so
we evaluate to Unset. Because of this, we don't need to deeply evaluate
pattern command arguments.