Change doc comments to use the dynamic binding returned by `doc:value` rather than the special `self` binding.
Simplify the dependencies of doc comment capture tracking. It's easier
and more consistent for the contents of doc comments to affect values
rather than the value not being affected by the doc comments whatsoever.
This might be changed back in the future, however in practice this
shouldn't cause any negative effects.