Remove the ValueByContent trait.
This removes the value_by_content function, the std:by-content function,
and changes tests to no longer rely on value_by_content when comparing
(which is admittedly a little cumbersome for nested value comparison,
but maybe in the future we'll make this more convenient).

Ensure a DependenciesConstant will calculate the same identity if
converted to a Dependencies (revealed by the removal of ValueByContent
with the std:bool tests).