Add an exceptional case for identity calculation for Command expressions.
If the command function is `eval_for_id` and all captures are ready, the
argument identities will not be computed. This is an optimization (as
the expression will be evaluated anyway) and may also allow certain
behaviors by making it possible to skip identity calculations for
certain values (not sure if this will be useful or not).