Add an `eval_for_id_hint` method to the data interface.
This not only works as an optimization (which seems to really speed
things up in practice), but importantly it allows the evaluator to
properly shortcut Command expressions. Previously, only Command
subexpressions could be shortcut due to the lack of semantic
differentiation between getting `eval_for_id` and the identity itself.

In the future, it may be appropriate to simply separate `eval_for_id`
and the identity rather than calculating and caching them together.