Refactor the way that the runtime and loading works.
The load cache is kept alive as long as the `load` function itself is
still reachable.

All created values reside in a parent lexical scope to greatly simplify
lifetime management, guaranteeing that a created value will be valid at
least as long as the parent function scope. All returned values are
added to the next ancestor scope.